contractguard-api/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.24.1
openai==1.91.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.4
supabase==2.3.3
pydantic-settings==2.5.2
azure-ai-documentintelligence==1.0.0
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-search==1.0.0b2
azure-search-documents==11.6.0b1
azure-storage-blob==12.19.0
openpyxl==3.1.5


